fix extra comma generation on import
the logic for extending an existing import statement failed to
recognize if the previous imported artifact ended with a comma if the
closing curly brace was in another line. for example:

import {
  A,
  B,
} from 'somewhere';

ts-index added an extra comma to the import statement above which made
the code syntactically invalid.

this patch avoids the extra comma.